# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meatballs API documentaion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Meatballs test server"
        },
        {
            "url": "https://api.meatballs.w6p.ru/api/",
            "description": "Meatballs production server"
        }
    ],
    "paths": {
        "/contacts/{role}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Get contact list by role",
                "description": "Retrieve a list of contacts by role",
                "operationId": "getContactsByRole",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "Role of contact to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Студент"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Студентов"
                                                    },
                                                    "middle_name": {
                                                        "type": "string",
                                                        "example": "Студентович"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "student@university.com"
                                                    },
                                                    "phone_number": {
                                                        "type": "string",
                                                        "example": "123-456-7890"
                                                    },
                                                    "img": {
                                                        "type": "string",
                                                        "example": null
                                                    },
                                                    "post": {
                                                        "properties": {
                                                            "post": {
                                                                "type": "string",
                                                                "example": "Староста"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "subject": {
                                                        "properties": {
                                                            "subject": {
                                                                "type": "string",
                                                                "example": "Технологии полиграфии"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "info": {
                                                        "properties": {
                                                            "faculty": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string",
                                                                    "example": "Информационная безопасность"
                                                                }
                                                            },
                                                            "education": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string",
                                                                    "example": "Очная"
                                                                }
                                                            },
                                                            "group": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string",
                                                                    "example": "ИБА-41"
                                                                }
                                                            },
                                                            "record_book": {
                                                                "type": "string",
                                                                "example": "1"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}